Controller Support and Bug Fixes
- add controller support
- fix bug where one missed ring causes no further rings to register
- add transparent ring around button to better show where to hit the incoming rings
