+++ Team1FinalProject/Assets/Scripts/IngredientDemo/NoteObject.cs
@@ -4,18 +4,12 @@ using UnityEngine;
 
 public class NoteObject : MonoBehaviour
 {
-    [SerializeField] private float _noteHitWidth = .3f;
-    [SerializeField] private float _goodNoteWidth = .2f;
-    [SerializeField] private float _perfectNoteWidth = .1f;
-    [SerializeField] private float _noteSize = 1f;
+    private float _noteHitWidth = .4f;
+    private float _goodNoteWidth = .2f;
+    private float _perfectNoteWidth = .1f;
+    private float _noteSize = 1f;
     public bool CanBePressed { get; private set; } = false;
-    public HitType HitType { get; private set; } = HitType.Missed;
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
+    public HitType HitType { get; private set; } = HitType.Upcoming;
 
     // Update is called once per frame
     void Update()
@@ -40,10 +34,14 @@ public class NoteObject : MonoBehaviour
                 HitType = HitType.Perfect;
             }
         }
-        else
+        else if (currentSize < lowerBound)
         {
             CanBePressed = false;
             HitType = HitType.Missed;
+        }
+        else
+        {
+            CanBePressed = false;
         }
     }
 }
