+++ Team1FinalProject/Assets/Scripts/Managers/NoteManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 21f0ce5c24d608e4fbdb68f1ee215153
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
